pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrivai"
version = "0.1.0"
description = "可配置的通用文档生成与审核框架"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "qmd>=0.1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
demo = [
    "streamlit>=1.43",
]

[tool.setuptools.packages.find]
include = ["core*", "utils*", "demo*"]

[tool.setuptools.package-data]
demo = ["config/*.yaml", "templates/*.j2"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "integration: 真实 API 调用测试（需要网络和 API key）",
    "unit: 单元测试（使用 mock）",
]
addopts = "-v --tb=short"